+++ packages/shopify-app-remix/src/server/authenticate/admin/billing/__tests__/request.test.ts
@@ -95,6 +95,7 @@ describe('Billing request', () => {
     // THEN
     expectExitIframeRedirect(response, {
       destination: responses.CONFIRMATION_URL,
+      addHostToExitIframePath: false,
     });
   });
 
@@ -166,7 +167,8 @@ describe('Billing request', () => {
 
   it('redirects to exit-iframe with authentication using app bridge when embedded and Shopify invalidated the session', async () => {
     // GIVEN
-    const shopify = shopifyApp(testConfig({billing: BILLING_CONFIG}));
+    const config = testConfig();
+    const shopify = shopifyApp({...config, billing: BILLING_CONFIG});
     await setUpValidSession(shopify.sessionStorage);
 
     await mockExternalRequest({
@@ -191,6 +193,9 @@ describe('Billing request', () => {
     );
 
     // THEN
+    const shopSessions =
+      await config.sessionStorage.findSessionsByShop(TEST_SHOP);
+    expect(shopSessions).toStrictEqual([]);
     expectExitIframeRedirect(response);
   });