+++ packages/shopify-app-remix/src/server/authenticate/admin/billing/cancel.ts
@@ -11,7 +11,7 @@ export function cancelBillingFactory(
   session: Session,
 ) {
   return async function cancelBilling(options: CancelBillingOptions) {
-    const {api, logger} = params;
+    const {api, logger, config} = params;
 
     logger.debug('Cancelling billing', {shop: session.shop, ...options});
 
@@ -27,6 +27,7 @@ export function cancelBillingFactory(
         logger.debug('API token was invalid, redirecting to OAuth', {
           shop: session.shop,
         });
+        await config.sessionStorage.deleteSession(session.id);
         throw await redirectToAuthPage(params, request, session.shop);
       } else {
         throw error;