+++ packages/shopify-app-remix/src/server/authenticate/admin/helpers/redirect.ts
@@ -66,15 +66,11 @@ function isBounceRequest(request: Request) {
 }
 
 function isDataRequest(request: Request) {
-  const {searchParams} = new URL(request.url);
-
   const isGet = request.method === 'GET';
   const sessionTokenHeader = Boolean(getSessionTokenHeader(request));
-  const sessionTokenSearchParam = searchParams.has('id_token');
 
   return (
     sessionTokenHeader &&
-    !sessionTokenSearchParam &&
     !isBounceRequest(request) &&
     (!isEmbeddedRequest(request) || !isGet)
   );