+++ packages/shopify-app-remix/src/server/authenticate/login/login.ts
@@ -35,7 +35,14 @@ export function loginFactory(params: BasicParams) {
       return {shop: LoginErrorType.InvalidShop};
     }
 
-    const redirectUrl = `${config.appUrl}${config.auth.path}?shop=${sanitizedShop}`;
+    const authPath = `${config.appUrl}${config.auth.path}?shop=${sanitizedShop}`;
+
+    const adminPath = api.utils.legacyUrlToShopAdminUrl(sanitizedShop);
+    const installPath = `https://${adminPath}/oauth/install?client_id=${config.apiKey}`;
+
+    const shouldInstall =
+      config.isEmbeddedApp && config.future.unstable_newEmbeddedAuthStrategy;
+    const redirectUrl = shouldInstall ? installPath : authPath;
 
     logger.info(`Redirecting login request to ${redirectUrl}`, {
       shop: sanitizedShop,