+++ packages/shopify-app-remix/src/server/errors.ts
@@ -1,0 +1,3 @@
+import {ShopifyError} from '@shopify/shopify-api';
+
+export class SessionNotFoundError extends ShopifyError {}