+++ packages/shopify-app-remix/src/server/future/flags.ts
@@ -14,6 +14,16 @@ export interface FutureFlags {
    * @default false
    */
   v3_authenticatePublic?: boolean;
+
+  /**
+   * When enabled, embedded apps will fetch access tokens via [token exchange](https://shopify.dev/docs/apps/auth/get-access-tokens/token-exchange).
+   * This assumes the app has scopes declared for [Shopify managing installation](https://shopify.dev/docs/apps/auth/installation#shopify-managed-installation).
+   *
+   * Learn more about this [new embedded app auth strategy](https://shopify.dev/docs/api/shopify-app-remix#embedded-auth-strategy).
+   *
+   * @default false
+   */
+  unstable_newEmbeddedAuthStrategy?: boolean;
 }
 
 export type FutureFlagOptions = FutureFlags | undefined;