+++ packages/shopify-app-remix/src/server/unauthenticated/admin/__tests__/factory.test.ts
@@ -1,4 +1,4 @@
-import {shopifyApp} from '../../../index';
+import {SessionNotFoundError, shopifyApp} from '../../../index';
 import {
   TEST_SHOP,
   setUpValidSession,
@@ -12,18 +12,18 @@ describe('unauthenticated admin context', () => {
     const shopify = shopifyApp(testConfig());
 
     // EXPECT
-    await expect(shopify.unauthenticated.admin(TEST_SHOP)).rejects.toThrow();
+    await expect(shopify.unauthenticated.admin(TEST_SHOP)).rejects.toThrow(
+      SessionNotFoundError,
+    );
   });
 
   expectAdminApiClient(async () => {
     const shopify = shopifyApp(testConfig());
-    const expectedSession = await setUpValidSession(
-      shopify.sessionStorage,
-      false,
-    );
-    const {admin, session: actualSession} = await shopify.unauthenticated.admin(
-      TEST_SHOP,
-    );
+    const expectedSession = await setUpValidSession(shopify.sessionStorage, {
+      isOnline: false,
+    });
+    const {admin, session: actualSession} =
+      await shopify.unauthenticated.admin(TEST_SHOP);
 
     return {admin, expectedSession, actualSession};
   });