+++ packages/shopify-app-remix/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @shopify/shopify-app-remix
 
+## 2.5.1
+
+### Patch Changes
+
+- 22b7744: Throw error when attempting to exit-iframe with an invalid URL
+
+## 2.5.0
+
+### Minor Changes
+
+- 6d5b4e0: Add check function to the billing API
+- ba3eff7: Added a new method `shopify.authenticate.flow(request)`, which will validate a Flow extension request, and return the payload / API clients to the app.
+- da09f4e: Add SessionNotFoundError for unauthenticated requests
+- 64e0246: Update shopify-api version to 9.2.0
+
+### Patch Changes
+
+- 8811faa: Updated the `@shopify/admin-api-client` dependency
+- f5742c1: Updated dependency on `@shopify/shopify-api`
+- d0a5483: Remove references to unstable_tokenExchange
+- 37dff44: Updated Remix dependencies to v2.5.1
+- Updated dependencies [f5742c1]
+- Updated dependencies [64e0246]
+  - @shopify/shopify-app-session-storage@2.1.0
+
+## 2.4.0
+
+### Minor Changes
+
+- 2473c85: Add new embedded authorization strategy relying on Shopify managed install and OAuth token exchange
+
+  :exclamation: For more information on how to enable this feature, see ["New Embedded Authorization Strategy"](./README.md#new-embedded-authorization-strategy)
+
+### Patch Changes
+
+- 35b74dd: Fixes a bug that was causing external redirects to fail in remix actions
+- 42013e8: Minor refactor in login.ts to use new URL util method from shopify-api-js
+- b4eeb24: Improved and simplified package.json dependencies
+- b998c30: Bump shopify-api version from 9.0.1 to 9.0.2
+- b998c30: Handle webhook registration throttling error
+- f445164: Use 'body' field from GraphqlQueryError when logging session validation error
+- Updated dependencies [b4eeb24]
+- Updated dependencies [b998c30]
+  - @shopify/shopify-app-session-storage@2.0.4
+
 ## 2.3.0
 
 ### Minor Changes